Add files via upload
Multithreading is a programming concept that involves the simultaneous execution of multiple threads in a single program or process. A thread is the smallest unit of execution within a process, and multithreading allows for concurrent execution of multiple threads to improve performance, responsiveness, and overall efficiency in certain types of applications.

The three mainly used concept in this project

Here are key concepts related to multithreading:

Thread:

A thread is the smallest unit of a process that can be independently scheduled and executed by the operating system.
Each thread has its own program counter, register set, and stack, but shares the same code section and data section with other threads in the same process.
Concurrency:

Concurrency refers to the ability of a program to execute multiple threads concurrently, allowing different parts of the program to make progress independently.
Parallelism:

Parallelism is a concept related to executing multiple tasks simultaneously by allocating each task to a separate processor or core. Multithreading is one way to achieve parallelism in a program